Fixed clang build

+++ source/base/math/ray.inl
@@ -13,9 +13,9 @@ inline Ray::Ray(const float3& origin, const float3& direction, float min_t, floa
 	min_t(min_t),
 	max_t(max_t),
 	signs {
-		inv_direction[0] < 0.f ? 1u : 0u,
-		inv_direction[1] < 0.f ? 1u : 0u,
-		inv_direction[2] < 0.f ? 1u : 0u}
+		inv_direction[0] < 0.f ? uint8_t(1u) : uint8_t(0u),
+		inv_direction[1] < 0.f ? uint8_t(1u) : uint8_t(0u),
+		inv_direction[2] < 0.f ? uint8_t(1u) : uint8_t(0u)}
 	{}
 
 inline Ray::Ray(const float3& origin, const float3& direction,
@@ -26,9 +26,9 @@ inline Ray::Ray(const float3& origin, const float3& direction,
 	min_t(min_t),
 	max_t(max_t),
 	signs {
-		inv_direction[0] < 0.f ? 1u : 0u,
-		inv_direction[1] < 0.f ? 1u : 0u,
-		inv_direction[2] < 0.f ? 1u : 0u},
+		inv_direction[0] < 0.f ? uint8_t(1u) : uint8_t(0u),
+		inv_direction[1] < 0.f ? uint8_t(1u) : uint8_t(0u),
+		inv_direction[2] < 0.f ? uint8_t(1u) : uint8_t(0u)},
 	depth(depth) {}
 
 inline Ray::Ray(const float3& origin, const float3& direction, const float3& inv_direction,
